Split PEG tests into peg (1-bag) and peg2 (2-bag, on-demand) and add candidate allowlist
- Extract shared test utilities (print, callback) into peg_test_util.h
- Move 2-bag test to peg2_test.c with assertions for 10I X(I) best move
- Add candidate_allowlist to PegArgs for targeted testing of specific plays
- Add 2peg_xi_extras_snot on-demand test: evaluates only X(I), EXT(R)AS, S(NO)T
