+++ test/peg2_test.h
@@ -1,0 +1,7 @@
+#ifndef PEG2_TEST_H
+#define PEG2_TEST_H
+
+void test_peg2(void);
+void test_peg2_xi_extras_snot(void);
+
+#endif
